refs #95 rpm/show action rework
*  needs name/pkgid and filters as input
*  does not use default values for filters, so that a query with only
   the name/pkgid parameter is likely to give results.
* if filters missing and given parameters not sufficient,
  disambiguation ("several RPMs match your selection, please choose one")
* from this action, going back to package or to other menu entries
  keeps the filters related to this RPM
** distrelease, source, media : kept
** media : not kept (in fact, given as temporary parameter t_media
   to the action)

+++ lib/filtering/criteriaFactory.class.php
@@ -2,7 +2,7 @@
 class criteriaFactory
 {
 
-  public function createFromContext(madbContext $context, $targetPerimeter)
+  public function createFromContext(madbContext $context, $targetPerimeter, $use_default_filtering_values = true)
   {
     $criteria              = new Criteria();
     $filtersIteratorFactory = new filtersIteratorFactory();
@@ -18,25 +18,25 @@ class criteriaFactory
       $perimeterFilters = $filterIterator->getByPerimeter($perimeter);
       if ($perimeter == $targetPerimeter)
       {
-        $criteria = $this->applyCurrentPerimeterFilters($perimeterFilters, $criteria, $context);
+        $criteria = $this->applyCurrentPerimeterFilters($perimeterFilters, $criteria, $context, $use_default_filtering_values);
       }
       else
       {
-        $criteria = $this->applyOtherPerimeterFilters($perimeterFilters, $criteria, $context, $perimeters->get($perimeter));
+        $criteria = $this->applyOtherPerimeterFilters($perimeterFilters, $criteria, $context, $perimeters->get($perimeter), $use_default_filtering_values);
       }
     }
 
     return $criteria;
   }
 
-  private function applyCurrentPerimeterFilters(filtersIterator $filters, $criteria, $context)
+  private function applyCurrentPerimeterFilters(filtersIterator $filters, $criteria, $context, $use_default_filtering_values)
   {
     $criteria = clone $criteria;
     foreach ($filters as $filter)
     {
       $filter->setCriteria($criteria);
       $filter->setMadbContext($context);
-      $criteria = $filter->getFilteredCriteria();
+      $criteria = $filter->getFilteredCriteria(true, $use_default_filtering_values);
     }
     return $criteria;
   }
@@ -46,15 +46,21 @@ class criteriaFactory
     return Propel::getConnection();
   }
 
-  private function applyOtherPerimeterFilters(filtersIterator $filters, Criteria $criteria, $context, basePerimeter $perimeter)
+  private function applyOtherPerimeterFilters(filtersIterator $filters, Criteria $criteria, $context, basePerimeter $perimeter, $use_default_filtering_values)
   {
     $criteriaOrig = clone $criteria;
-    $criteria     = new Criteria();
-    $criteria->setDistinct();
-    $criteria = $perimeter->addTemporayTableColumns($criteria);
-    $criteria = $this->applyCurrentPerimeterFilters($filters, $criteria, $context);
+    $selectall_criteria = new Criteria();
+    $selectall_criteria->setDistinct();
+    $selectall_criteria = $perimeter->addTemporayTableColumns($selectall_criteria);
+    
+    $criteria = $this->applyCurrentPerimeterFilters($filters, $selectall_criteria, $context, $use_default_filtering_values);
 
-
+    // No need to create a temp table which contains every package or rpm id
+    if ($criteria == $selectall_criteria)
+    {
+      return $criteriaOrig;
+    }
+    
     $tablename = 'tmp_filtrage_' . md5(serialize($filters));//TODO better filtertablename
     //TODO do not delete every time this table
     $this->getConnection()->exec(sprintf('DROP TABLE IF EXISTS `%s`', $tablename));
@@ -67,9 +73,9 @@ class criteriaFactory
 
     $criteria = $criteriaOrig;
 
-    //TODO by perumeter join
+    //TODO by perimeter join
 
-    //deux mthodes getTargetId ?
+    //2 getTargetId methods?
     if (get_class($perimeter) != 'rpmPerimeter')
     {
       $criteria->addJoin(RpmPeer::PACKAGE_ID, $toTmp->getField('id'), Criteria::JOIN);
